Added Perl NVRAM block editor
This commit adds a demonstration editor for the NVRAM block generated by
"nvram.c". The editor script, "editor.pl", uses doxygen to parse the C file and
generate and XML file which the Perl script can use to determine the contents
of the NVRAM block.

+++ nvram.c
@@ -150,9 +150,9 @@ static int block(char *buffer, size_t length, const char *name, bool read)
 	r = read ? fread(buffer, 1, length, l) : fwrite(buffer, 1, length, l);
 	if (r != length) {
 		fprintf(stderr, "warning - partial block operation "
-				" (%zu/%zu bytes %s): %s\n", 
-				r, 
-				length, 
+				" (%u/%u bytes %s): %s\n", 
+				(unsigned)r, 
+				(unsigned)length, 
 				read ? "read" : "wrote", 
 				strerror(errno));
 		fclose(l);
